+++ partest/project_gen/emitters/__init__.py
@@ -1,5 +1,0 @@
-"""Code emitters for project_gen (G2+)."""
-
-from partest.project_gen.emitters.resources import emit_resources
-
-__all__ = ["emit_resources"]